# Fix performance regression of x86_64 GCM with rustc 1.87